+++ CHANGES.rst
@@ -1,5 +1,10 @@
 Changes
 -------
+2.22.0 (2025-04-07)
+^^^^^^^^^^^^^^^^^^^
+* upstream support for the smithy-rpc-v2-cbor protocol
+* bump botocore dependency specification
+
 2.21.1 (2025-03-04)
 ^^^^^^^^^^^^^^^^^^^
 * fix for refreshable credential account-id lookup